koda-sandbox: split proxy.rs into proxy/{env,external,handle}.rs (Phase 3b prep, #934)
Phase 3b slice 1/N: pure mechanical refactor, zero behavior change.
All 174 tests pass byte-identically pre/post.

## Why split now

proxy.rs was 447 lines after 3a. 3b will add proxy/filter.rs (~80 LOC),
proxy/server.rs (~300 LOC), proxy/builtin.rs (~150 LOC) — keeping
everything in one file would push it well past 600 lines and make each
new feature commit mix mechanical concerns with new code. Splitting
once, up front, keeps subsequent diffs focused.

## Layout

proxy/
├── mod.rs       — top-level docstring + pub(crate) helpers
│                 (pick_ephemeral_port, wait_for_bind) shared between
│                 every spawn path (today: ExternalProxy; 3b:
│                 BuiltInProxy)
├── env.rs       — DEFAULT_NO_PROXY, PROXY_PORT_ENV_KEY,
│                 proxy_env_vars, ca_bundle_for_policy + their tests
├── external.rs  — ExternalProxy + 3 unit tests
└── handle.rs    — ProxyHandle (refactored; see below)

## Refactor: ProxyHandle inner enum

Old shape: ProxyHandle had a public "port" field and a private
"Option<Child>". Worked fine for one variant but doesn't generalize
to the 3b built-in proxy whose lifecycle is a tokio task, not a child
process.

New shape: ProxyHandle holds an Inner enum:

  enum Inner {
      External(Option<Child>),
      // 3b: BuiltIn(JoinHandle<()>),
  }

Public surface unchanged — port (still a public field), ca_bundle()
still returns Option<&Path> (always None for External), shutdown()
still idempotent, Drop still calls shutdown(). All existing callers
keep working without any signature changes.

Constructor is now ProxyHandle::from_child(port, child),
pub(crate)-only — callers reach it through ExternalProxy::spawn (3a)
or BuiltInProxy::spawn (3b).

## Re-exports unchanged

pub use proxy::{DEFAULT_NO_PROXY, ExternalProxy, PROXY_PORT_ENV_KEY,
ProxyHandle, ca_bundle_for_policy, proxy_env_vars} still works at
the crate root. WorkerClient::spawn_with_policy_and_proxy and the
Phase 3a integration tests didn't need a single line changed.

## Tests: 174 passing

152 unit + 16 integration + 3 snapshot + 2 worker + 1 doc — clippy
clean, fmt clean, rustdoc clean.
