+++ koda-sandbox/src/proxy/env.rs
@@ -1,0 +1,184 @@
+//! Env-var bouquet for routing sandboxed subprocesses through the proxy
+//! (Phase 3a of #934).
+//!
+//! See [parent module docs](super) for the why.
+
+use std::path::Path;
+
+/// Default `NO_PROXY` value: loopback + RFC1918 + AWS/GCE IMDS.
+///
+/// Borrowed verbatim from Claude Code's `upstreamproxy.ts` `NO_PROXY_LIST`
+/// (which itself mirrors the Bun/curl/Go/Python intersection of supported
+/// patterns). These are the addresses every reasonable proxy declines to
+/// intercept — without them, sandboxed processes can't talk to localhost
+/// dev servers, can't read instance metadata, and can't reach RFC1918
+/// services on the user's LAN.
+pub const DEFAULT_NO_PROXY: &str =
+    "localhost,127.0.0.1,::1,169.254.0.0/16,10.0.0.0/8,172.16.0.0/12,192.168.0.0/16";
+
+/// Env-var name passed to a user proxy command so it knows which port to bind.
+///
+/// Mirrors Codex's `PROXY_ACTIVE_ENV_KEY` pattern. Avoids template-string
+/// parsing and lets the proxy command be a plain `Vec<String>`.
+pub const PROXY_PORT_ENV_KEY: &str = "KODA_PROXY_PORT";
+
+/// Generate the env-var bouquet for a sandboxed subprocess.
+///
+/// `port` is where the proxy is listening on `127.0.0.1`. `ca_bundle` is
+/// the path to a PEM bundle the subprocess should trust for TLS verification
+/// — typically points to a corporate CA + system CA concatenation. When
+/// `None`, the cert-bundle vars are omitted (the subprocess uses its
+/// platform default trust store).
+///
+/// Returned as `Vec<(String, String)>` so the caller can `.envs(...)` it
+/// directly into a `Command` builder. Sorted by key for deterministic
+/// snapshot tests.
+///
+/// ## Why so many keys?
+///
+/// Different runtimes look at different env vars:
+///
+/// | Runtime | Proxy var | CA bundle var |
+/// |---|---|---|
+/// | curl, libcurl | `HTTPS_PROXY` (UPPER) | `CURL_CA_BUNDLE` |
+/// | Python `requests` | `HTTPS_PROXY` (UPPER) | `REQUESTS_CA_BUNDLE` |
+/// | Python `httpx`, `urllib` | `https_proxy` (lower) | `SSL_CERT_FILE` |
+/// | Node.js (undici) | `HTTPS_PROXY` (UPPER) | `NODE_EXTRA_CA_CERTS` |
+/// | Go (`net/http`) | `HTTPS_PROXY` (UPPER) | `SSL_CERT_FILE` |
+/// | Rust (`reqwest`) | `HTTPS_PROXY` (UPPER) | `SSL_CERT_FILE` |
+///
+/// Setting all of them is the only way to cover every dev tool without
+/// per-tool wrappers.
+pub fn proxy_env_vars(port: u16, ca_bundle: Option<&Path>) -> Vec<(String, String)> {
+    let proxy_url = format!("http://127.0.0.1:{port}");
+
+    let mut vars = vec![
+        ("HTTPS_PROXY".to_string(), proxy_url.clone()),
+        ("https_proxy".to_string(), proxy_url.clone()),
+        ("HTTP_PROXY".to_string(), proxy_url.clone()),
+        ("http_proxy".to_string(), proxy_url),
+        ("NO_PROXY".to_string(), DEFAULT_NO_PROXY.to_string()),
+        ("no_proxy".to_string(), DEFAULT_NO_PROXY.to_string()),
+    ];
+
+    if let Some(ca) = ca_bundle {
+        let ca_str = ca.to_string_lossy().to_string();
+        vars.push(("SSL_CERT_FILE".to_string(), ca_str.clone()));
+        vars.push(("NODE_EXTRA_CA_CERTS".to_string(), ca_str.clone()));
+        vars.push(("REQUESTS_CA_BUNDLE".to_string(), ca_str.clone()));
+        vars.push(("CURL_CA_BUNDLE".to_string(), ca_str));
+    }
+
+    // Deterministic order for snapshot tests.
+    vars.sort_by(|a, b| a.0.cmp(&b.0));
+    vars
+}
+
+/// Path to the CA bundle to advertise via env vars, given a [`crate::policy::NetPolicy`].
+///
+/// Returns `None` when no MITM is configured — in that case the subprocess
+/// uses its platform default trust store. Returning `Option<&Path>` rather
+/// than `Option<PathBuf>` so callers can pass it directly to [`proxy_env_vars`]
+/// without intermediate allocation.
+pub fn ca_bundle_for_policy(net: &crate::policy::NetPolicy) -> Option<&Path> {
+    net.mitm.as_ref().map(|m| m.ca_bundle.as_path())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::path::PathBuf;
+
+    #[test]
+    fn proxy_env_vars_includes_all_six_proxy_keys() {
+        let vars = proxy_env_vars(8080, None);
+        let keys: Vec<&str> = vars.iter().map(|(k, _)| k.as_str()).collect();
+
+        assert!(keys.contains(&"HTTPS_PROXY"));
+        assert!(keys.contains(&"https_proxy"));
+        assert!(keys.contains(&"HTTP_PROXY"));
+        assert!(keys.contains(&"http_proxy"));
+        assert!(keys.contains(&"NO_PROXY"));
+        assert!(keys.contains(&"no_proxy"));
+    }
+
+    #[test]
+    fn proxy_env_vars_omits_ca_bundle_when_none() {
+        let vars = proxy_env_vars(8080, None);
+        let keys: Vec<&str> = vars.iter().map(|(k, _)| k.as_str()).collect();
+
+        assert!(!keys.contains(&"SSL_CERT_FILE"));
+        assert!(!keys.contains(&"NODE_EXTRA_CA_CERTS"));
+        assert!(!keys.contains(&"REQUESTS_CA_BUNDLE"));
+        assert!(!keys.contains(&"CURL_CA_BUNDLE"));
+    }
+
+    #[test]
+    fn proxy_env_vars_includes_all_four_ca_keys_when_some() {
+        let bundle = PathBuf::from("/etc/ssl/corp-ca.pem");
+        let vars = proxy_env_vars(8080, Some(&bundle));
+        let keys: Vec<&str> = vars.iter().map(|(k, _)| k.as_str()).collect();
+
+        assert!(keys.contains(&"SSL_CERT_FILE"));
+        assert!(keys.contains(&"NODE_EXTRA_CA_CERTS"));
+        assert!(keys.contains(&"REQUESTS_CA_BUNDLE"));
+        assert!(keys.contains(&"CURL_CA_BUNDLE"));
+
+        // All four point at the same path.
+        for key in [
+            "SSL_CERT_FILE",
+            "NODE_EXTRA_CA_CERTS",
+            "REQUESTS_CA_BUNDLE",
+            "CURL_CA_BUNDLE",
+        ] {
+            let v = vars
+                .iter()
+                .find(|(k, _)| k == key)
+                .map(|(_, v)| v.as_str())
+                .unwrap();
+            assert_eq!(v, "/etc/ssl/corp-ca.pem");
+        }
+    }
+
+    #[test]
+    fn proxy_url_format_uses_loopback_ipv4() {
+        let vars = proxy_env_vars(31415, None);
+        let url = vars
+            .iter()
+            .find(|(k, _)| k == "HTTPS_PROXY")
+            .map(|(_, v)| v.as_str())
+            .unwrap();
+        assert_eq!(url, "http://127.0.0.1:31415");
+    }
+
+    #[test]
+    fn no_proxy_default_covers_loopback_and_rfc1918() {
+        // Sanity check that we haven't accidentally truncated the constant.
+        assert!(DEFAULT_NO_PROXY.contains("127.0.0.1"));
+        assert!(DEFAULT_NO_PROXY.contains("::1"));
+        assert!(DEFAULT_NO_PROXY.contains("10.0.0.0/8"));
+        assert!(DEFAULT_NO_PROXY.contains("172.16.0.0/12"));
+        assert!(DEFAULT_NO_PROXY.contains("192.168.0.0/16"));
+        // AWS / GCE IMDS link-local: dropping this would prevent cloud
+        // workloads from reading instance metadata.
+        assert!(DEFAULT_NO_PROXY.contains("169.254.0.0/16"));
+    }
+
+    #[test]
+    fn ca_bundle_for_policy_handles_no_mitm() {
+        let policy = crate::policy::NetPolicy::default();
+        assert!(ca_bundle_for_policy(&policy).is_none());
+    }
+
+    #[test]
+    fn ca_bundle_for_policy_returns_path_when_mitm_set() {
+        let policy = crate::policy::NetPolicy {
+            mitm: Some(crate::policy::MitmConfig {
+                ca_bundle: PathBuf::from("/x/ca.pem"),
+                socket_map: vec![],
+            }),
+            ..Default::default()
+        };
+        assert_eq!(ca_bundle_for_policy(&policy), Some(Path::new("/x/ca.pem")));
+    }
+}
